+++ toot/tui/widgets.py
@@ -27,3 +27,22 @@ class SelectableText(Clickable, urwid.Text):
 
 class SelectableColumns(Clickable, urwid.Columns):
     _selectable = True
+
+
+class EditBox(urwid.AttrWrap):
+    """Styled edit box."""
+    def __init__(self):
+        edit = urwid.Edit(multiline=True, allow_tab=True)
+        return super().__init__(edit, "editbox", "editbox_focused")
+
+
+class Button(urwid.AttrWrap):
+    """Styled button."""
+    def __init__(self, *args, **kwargs):
+        button = urwid.Button(*args, **kwargs)
+        padding = urwid.Padding(button, width=len(args[0]) + 4)
+        return super().__init__(padding, "button", "button_focused")
+
+    def set_label(self, *args, **kwargs):
+        self.original_widget.original_widget.set_label(*args, **kwargs)
+        self.original_widget.width = len(args[0]) + 4
